fix(#37): Add legal notices for intellij-ce and notices about modified files
