+++ javalint-formatter/src/main/java/com/intellij/psi/impl/LocalImpl.java
@@ -10,7 +10,15 @@ import com.intellij.psi.PsiFile;
 import com.intellij.psi.util.PsiUtilCore;
 import org.jetbrains.annotations.NotNull;
 
+/**
+ * This file has been modified from its original version.
+ * Modifications made by Kamil Perczyński on 2024-01-14
+ * <p>
+ * Modifications:
+ * <li>Removed check for synchronization between Document and PsiFile
+ */
 public class LocalImpl extends PsiDocumentManagerBase {
+
   protected LocalImpl(@NotNull Project project) {
     super(project);
   }
@@ -34,15 +42,6 @@ public class LocalImpl extends PsiDocumentManagerBase {
     VirtualFile virtualFile = viewProvider.getVirtualFile();
     document = FileDocumentManager.getInstance().getDocument(virtualFile, myProject);
     if (document != null) {
-//      if (document.getTextLength() != file.getTextLength()) {
-//        // We have internal state inconsistency, it might be a good idea to contact the core team if you are able to reproduce this error.
-//        String message = "Document/PSI mismatch: " + file + " of " + file.getClass() +
-//          "; viewProvider=" + viewProvider +
-//          "; uncommitted=" + Arrays.toString(getUncommittedDocuments());
-//
-//        System.out.println("WARN: " + message);
-//      }
-
       if (!viewProvider.isPhysical()) {
         PsiUtilCore.ensureValid(file);
         associatePsi(document, file);
@@ -52,4 +51,5 @@ public class LocalImpl extends PsiDocumentManagerBase {
 
     return document;
   }
+
 }
