+++ javalint-formatter/src/main/resources/META-INF/NOTICE
@@ -1,0 +1,3 @@
+This software includes code from IntelliJ IDEA Community Edition
+Copyright (C) JetBrains s.r.o.
+https://www.jetbrains.com/idea/
